fini

+++ model/profil.php
@@ -5,12 +5,12 @@
 	function add_file(){
 		$bool = false;
 		if(isset($_POST['sumbit_add_file'])){
-			if(isset($_FILES['file']['name']) && !empty($_FILES)){
-				$file_tmp_name = $_FILES['file']['tmp_name'];
+			if(isset($_FILES['add_file']['name']) && !empty($_FILES)){
+				$file_tmp_name = $_FILES['add_file']['tmp_name'];
 				$directory = $_POST['directory'];
 				$default_choice = 'ajouter_dans_un_dossier';
 				if($_POST['nickname'] == ''){
-					$file['file_name'] = $_FILES['file']['name'];
+					$file['file_name'] = $_FILES['add_file']['name'];
 					if($directory != $default_choice){
 						$file['url'] = 'uploads/'.$_SESSION['user_username'] . '/' . $directory .'/' . $file["file_name"];
 					}else{
@@ -21,13 +21,13 @@
 
 	 				if(extension_accept($file['file_name'])){
 	 					if(!file_exist($file['url'])){
-		 					db_insert('files', $file);
-		 					move_uploaded_file($file_tmp_name, $file['url']);
-		 					$bool = true;
+			 				db_insert('files', $file);
+			 				move_uploaded_file($file_tmp_name, $file['url']);
+			 				$bool = true;
 	 					}
 	 				}	
 				}else{
-					$name_die = $_FILES['file']['name'];
+					$name_die = $_FILES['add_file']['name'];
 					$extension = file_extension($name_die);
 					$file['file_name'] = $_POST['nickname'].$extension;
 					if($directory != $default_choice){
@@ -98,7 +98,7 @@
 	}
 
 	function extension_accept($file_name){
-		$extension_autorisees = array('.jpg', '.jpeg', '.txt','.png','.pdf');
+		$extension_autorisees = array('.jpg', '.jpeg', '.txt','.png','.pdf', '.mp3', '.mp4');
 		$format = file_extension($file_name);
 		return in_array($format, $extension_autorisees);
 	}
@@ -157,27 +157,51 @@
 		$bool = false;
 		if(isset($_POST['submit_replace_file'])){
 			if(isset($_FILES['file']['name']) && !empty($_FILES) 
-				&& $_POST['file_name_to_replace'] !== ''){
+				&& $_POST['file_name_to_replace'] !== '' && 
+				$_POST['directory_replace'] !== ''){
 				$id_user = $_SESSION['user_id'];
 				$file_tmp_name = $_FILES['file']['tmp_name'];
 				$file_name_to_replace = $_POST['file_name_to_replace'];
 				$new_file_name = $_FILES['file']['name'];
-				if(file_exist_by_name($file_name_to_replace) && 
+				$directory = $_POST['directory_replace'];
+				$default_choice = 'Racine';
+				if($directory == $default_choice){
+					if(file_exist_by_name($file_name_to_replace) && 
 					$new_file_name !== '' ){
-					if(!file_exist_by_name($new_file_name)){
-						$data = get_file_by_file_name($file_name_to_replace);
-						$file_url = $data['file_url'];
-						$new_file_url = substr($file_url, 0, -(strlen($file_name_to_replace))).$new_file_name;
-						if(!find_one_secure("UPDATE files SET file_name = :new_file_name , file_url = :new_file_url  WHERE id_user = :id_user AND file_name = :file_name_to_replace",
-                            ['new_file_name' => $new_file_name,
-                             'new_file_url' => $new_file_url,
-                             'file_name_to_replace' => $file_name_to_replace,
-                             'id_user' => $id_user])){
-							move_uploaded_file($file_tmp_name,$new_file_url);
-                            unlink($file_url);
-                            $bool = true;
+						if(!file_exist_by_name($new_file_name)){
+							$data = get_file_by_file_name($file_name_to_replace);
+							$file_url = $data['file_url'];
+							$new_file_url = substr($file_url, 0, -(strlen($file_name_to_replace))).$new_file_name;
+							echo $new_file_url;
+							if(!find_one_secure("UPDATE files SET file_name = :new_file_name , file_url = :new_file_url  WHERE id_user = :id_user AND file_name = :file_name_to_replace",
+	                            ['new_file_name' => $new_file_name,
+	                             'new_file_url' => $new_file_url,
+	                             'file_name_to_replace' => $file_name_to_replace,
+	                             'id_user' => $id_user])){
+								move_uploaded_file($file_tmp_name,$new_file_url);
+	                            unlink($file_url);
+	                            $bool = true;
+							}
 						}
-
+					}
+				}else{
+					$data = get_file_by_file_name($file_name_to_replace);
+					$file_url = $data['file_url'];
+					$new_file_url = substr($file_url, 0, -(strlen($file_name_to_replace))).$new_file_name;
+					echo 'old ' .$file_url.'<br>';
+					echo 'new url ' .$new_file_url;
+					if(!file_exist($new_file_url)){
+						if(!find_one_secure("UPDATE files SET file_name = :new_file_name , file_url = :new_file_url  WHERE id_user = :id_user AND file_url = :file_url",
+	                            ['new_file_name' => $new_file_name,
+	                             'new_file_url' => $new_file_url,
+	                             'file_url' => $file_url,
+	                             'id_user' => $id_user])){
+								echo 'old' .$file_url.'<br>';
+								echo $new_file_url;
+								move_uploaded_file($file_tmp_name,$new_file_url);
+	                            unlink($file_url);
+	                            $bool = true;
+							}
 					}
 				}
 			}
@@ -273,6 +297,91 @@
 		}
 		return $bool;
 	}
+
+	function rename_folder(){
+		$bool = false;
+		if(isset($_POST['submit_rename_folder'])){
+			if(isset($_POST['new_folder_name']) && $_POST['new_folder_name'] != '' &&
+				isset($_POST['old_folder_name']) && $_POST['old_folder_name'] != ''){
+				$id_user = $_SESSION['user_id'];
+				$dir = 'uploads/'.$_SESSION['user_username'];
+				$all_dir = dirToArray($dir);
+				$old_folder_name = $_POST['old_folder_name'];
+				$new_folder_name = $_POST['new_folder_name'];
+				foreach ($all_dir as $key => $value) {
+					if($key == $old_folder_name){
+						if(is_array($value)){
+							if(!empty($value)){
+								foreach ($value as $key2 => $value2) {
+									$file_url = $dir.'/'.$key.'/'.$value2;
+									$new_file_url = $dir.'/'.$new_folder_name.'/'.$value2;
+									
+									if(!find_one_secure("UPDATE files SET file_url = :new_file_url  WHERE id_user = :id_user AND file_url = :file_url",
+	                            	['new_file_url' => $new_file_url,
+	                             	'file_url' => $file_url,
+	                             	'id_user' => $id_user])){
+										rename($dir.'/'.$key, $dir.'/'.$new_folder_name);
+										$bool = true;
+									}
+							
+								}
+							
+							}else{
+								rename($dir.'/'.$key, $dir.'/'.$new_folder_name);
+							}
+						}
+					}
+				}
+			}
+		}
+		return $bool;
+	}
+
+	/*function zipMyFolder(){
+
+    if(isset($_POST['downloadDirectory'])){
+    // Get real path for our folder
+$rootPath = $_POST["directoryToDownload"];
+
+// Initialize archive object
+$zip = new ZipArchive();
+$archive_file_name = basename($_POST["directoryToDownload"]).'.zip';
+$zip->open($_POST["directoryToDownload"].'.zip', ZipArchive::CREATE | ZipArchive::OVERWRITE);
+
+$files = new RecursiveIteratorIterator(
+    new RecursiveDirectoryIterator($rootPath),
+    RecursiveIteratorIterator::LEAVES_ONLY
+);
+
+foreach ($files as $name => $file)
+{
+
+    if (!$file->isDir())
+    {
+
+        $filePath = $file->getRealPath();
+        $relativePath = substr($filePath, strlen($rootPath) + 1);
+        $zip->addFile($filePath, $relativePath);
+    }
+}
+
+$zip->close();
+$user = get_user_by_id($_SESSION['user_id']);
+$username = $user["lastname"];
+header("Content-type: application/zip"); 
+header("Content-Disposition: attachment; filename=$archive_file_name");
+header("Content-length: " . filesize("uploads/".$username.'/'.$archive_file_name));
+header("Pragma: no-cache"); 
+header("Expires: 0"); 
+readfile("uploads/"."$username"."/"."$archive_file_name");
+    }
+}
+
+            <form action='?action=profil' method='post' class='formDirectory' name='formDownloadDirectory'><label for='downloadMyDir'>
+            <img class='imgDirect' src='assets/img/downloadfolder.png'></label>
+            <input type='text' name='directoryToDownload' value=".realpath($allmydir)." class='noneInputText'>
+            <input type='submit' name='downloadDirectory' value=".realpath($allmydir)." class='noneMyDelete' id='downloadMyDir'>
+            </form>*/
 
 
 ?>
